backend/requirement.txt
aiodns==3.2.0
aiofile==3.9.0
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.3.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
art==6.0
asgiref==3.7.2
asttokens==2.4.1
attrs==23.2.0
autopep8==1.4.4
beautifulsoup4==4.12.3
bleach==6.1.0
Brotli==1.1.0
cachetools==5.5.2
caio==0.9.21
certifi==2019.3.9
cffi==1.16.0
chardet==3.0.4
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
comm==0.2.1
contourpy==1.2.1
cryptography==44.0.3
cycler==0.12.1
datasets==3.2.0
debugpy==1.8.1
decorator==5.1.1
defusedxml==0.6.0
dill==0.3.8
distlib==0.3.9
Django==5.0.8
django-admin-thumbnails==0.2.9
django-allauth==0.50.0
django-cors-headers==4.6.0
django-countries==5.3.3
django-crispy-forms==1.7.2
django-debug-toolbar==1.11.1
django-extensions==3.2.3
django-jazzmin==3.0.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
executing==2.0.1
fastjsonschema==2.19.1
filelock==3.16.1
fonttools==4.51.0
fqdn==1.5.1
frozenlist==1.5.0
fsspec==2024.9.0
google-ai-generativelanguage==0.6.15
google-api-core==2.24.2
google-api-python-client==2.167.0
google-auth==2.39.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.5
googleapis-common-protos==1.70.0
graphviz==0.20.3
grpcio==1.71.0
grpcio-status==1.71.0
gunicorn==23.0.0
httplib2==0.22.0
huggingface-hub==0.27.0
idna==2.8
ipykernel==6.29.3
ipython==8.22.1
isoduration==20.11.0
jedi==0.19.1
Jinja2==3.1.3
joblib==1.4.2
jsonpointer==2.4
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter-events==0.9.1
jupyter_client==8.6.0
jupyter_core==5.7.1
jupyter_server==2.13.0
jupyter_server_terminals==0.5.3
jupyterlab_pygments==0.3.0
kiwisolver==1.4.5
MarkupSafe==2.1.5
matplotlib==3.8.4
matplotlib-inline==0.1.6
mistune==3.0.2
multidict==6.1.0
multiprocess==0.70.16
mutagen==1.47.0
mysqlclient==2.2.4
nbclient==0.10.0
nbconvert==7.16.2
nbformat==5.10.3
nest-asyncio==1.6.0
nltk==3.9.1
numpy==1.26.4
oauthlib==3.0.1
overrides==7.7.0
packaging==23.2
pandas==2.2.2
pandocfilters==1.5.1
parso==0.8.3
pep8==1.7.1
pillow==10.2.0
platformdirs==4.2.0
prometheus_client==0.20.0
prompt-toolkit==3.0.43
propcache==0.2.1
proto-plus==1.26.1
protobuf==5.29.4
psutil==5.9.8
pure-eval==0.2.2
pyarrow==18.1.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycares==4.5.0
pycodestyle==2.5.0
pycparser==2.21
pycryptodomex==3.20.0
pydantic==2.11.3
pydantic_core==2.33.1
Pygments==2.17.2
PyJWT==2.10.1
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-decouple==3.1
python-dotenv==1.1.0
python-json-logger==2.0.7
python3-openid==3.1.0
pytz==2024.1
pywin32==306
pywinpty==2.0.13
PyYAML==6.0.1
pyzmq==25.1.2
referencing==0.34.0
regex==2024.11.6
requests==2.32.3
requests-oauthlib==1.2.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.18.0
rsa==4.9.1
scikit-learn==1.5.2
scipy==1.14.1
seaborn==0.13.2
Send2Trash==1.8.2
six==1.16.0
sniffio==1.3.1
soupsieve==2.5
sqlparse==0.5.1
stack-data==0.6.3
stream-chat==4.23.0
stripe==2.27.0
tabulate==0.9.0
terminado==0.18.1
threadpoolctl==3.5.0
tinycss2==1.2.1
tornado==6.4
tqdm==4.67.1
traitlets==5.14.1
types-python-dateutil==2.9.0.20240316
typing-inspection==0.4.0
typing_extensions==4.12.2
tzdata==2024.1
uri-template==1.3.0
uritemplate==4.1.1
urllib3==2.2.2
virtualenv==20.28.0
wcwidth==0.2.13
webcolors==1.13
webencodings==0.5.1
websocket-client==1.7.0
websockets==13.0.1
whitenoise==6.7.0
xxhash==3.5.0
yarl==1.18.3
yt-dlp==2024.8.6
